[fix] token issue and user deleted issues, and blog creating isues

+++ src/app/modules/blog/blog.service.ts
@@ -4,9 +4,14 @@ import mongoose from 'mongoose';
 import { BlogSearchableFields } from './blog.constant';
 import { BlogQueryParams, TBlog } from './blog.interface';
 import { BlogModel } from './blog.model';
+import AppError from '../../errors/appError';
+import { StatusCodes } from 'http-status-codes';
 
-const createBlogIntoDB = async (payload: TBlog) => {
+const createBlogIntoDB = async (user: any, payload: TBlog) => {
+  payload.author = user._id;
+
   const result = await BlogModel.create(payload);
+
   return result.populate('author');
 };
 
@@ -52,8 +57,19 @@ const updateBlogFromDB = async (id: string, payload: Partial<TBlog>) => {
   return updateBlog.populate('author');
 };
 
-const deleteBlogFromDB = async (id: string) => {
-  await BlogModel.findByIdAndDelete(id);
+const deleteBlogFromDB = async (userId: string, id: string) => {
+  const blog = await BlogModel.findById(id).populate('author');
+  if (!blog) {
+    throw new AppError(StatusCodes.NOT_FOUND, 'Blog not found!');
+  }
+
+  const blogAuthorId = blog.author?._id;
+
+  if (blogAuthorId && blogAuthorId.toString() == userId) {
+    await BlogModel.findByIdAndDelete(id);
+  } else {
+    throw new AppError(StatusCodes.BAD_REQUEST, 'You did not deleted the blog');
+  }
 };
 
 export const BlogServices = {
